eve-server/requirements.txt
# --- 核心框架 ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# --- 数据库与异步驱动 ---
# SQLAlchemy 2.0 异步支持的核心
sqlalchemy>=2.0.27
# PostgreSQL 的异步驱动
asyncpg>=0.29.0
# 数据库迁移工具
alembic>=1.13.1
# 用于解析 PostgreSQL 连接串
psycopg2-binary>=2.9.9

# --- 认证与安全 ---
# 处理 ESI 的 JWT Token
python-jose[cryptography]>=3.3.0
# 密码哈希（虽然目前主推 ESI 登录，但建议预留）
passlib[bcrypt]>=1.7.4
# 处理 OAuth2 表单
python-multipart>=0.0.9

# --- 网络请求 ---
# 异步请求 ESI 接口必备，替代 requests
httpx>=0.27.0
aiohttp>=3.8.0
# --- 工具类 ---
# 加密敏感数据
cryptography>=42.0.0
# 环境变量加载
python-dotenv>=1.0.1